+++ example/lib/home.dart
@@ -20,8 +20,8 @@ class Home extends StatelessWidget {
             ),
             ElevatedButton(
               onPressed: () {
-                SupabaseAuthUi().signOutActiveUser();
-                Navigator.of(context).popUntil(ModalRoute.withName('/'));
+                Supabase.instance.client.auth.signOut();
+                Navigator.of(context).pushReplacementNamed('/');
               },
               child: const Text(
                 'Log Out',