+++ lib/src/components/supa_magic_auth.dart
@@ -1,17 +1,27 @@
 import 'package:email_validator/email_validator.dart';
 import 'package:flutter/material.dart';
 import 'package:supabase_auth_ui/src/utils/constants.dart';
-import 'package:supabase_auth_ui/src/utils/supabase_auth_ui.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
+/// UI component to create magic link login form
 class SupaMagicAuth extends StatefulWidget {
+  /// `redirectUrl` to be passed to the `.signIn()` or `signUp()` methods
+  ///
+  /// Typically used to pass a DeepLink
   final String? redirectUrl;
-  final void Function(GotrueSessionResponse response)? onSuccess;
-  final bool Function(GoTrueException error)? onError;
 
-  const SupaMagicAuth(
-      {Key? key, this.redirectUrl, this.onSuccess, this.onError})
-      : super(key: key);
+  /// Method to be called when the auth action is success
+  final void Function(Session response) onSuccess;
+
+  /// Method to be called when the auth action threw an excepction
+  final void Function(Object error)? onError;
+
+  const SupaMagicAuth({
+    Key? key,
+    this.redirectUrl,
+    required this.onSuccess,
+    this.onError,
+  }) : super(key: key);
 
   @override
   State<SupaMagicAuth> createState() => _SupaMagicAuthState();
@@ -20,21 +30,31 @@ class SupaMagicAuth extends StatefulWidget {
 class _SupaMagicAuthState extends State<SupaMagicAuth> {
   final _formKey = GlobalKey<FormState>();
   final _email = TextEditingController();
+  late final GotrueSubscription _gotrueSubscription;
 
-  final _supaAuth = SupabaseAuthUi();
+  bool _isLoading = false;
 
-  bool isLoading = false;
+  @override
+  void initState() {
+    super.initState();
+    _gotrueSubscription =
+        Supabase.instance.client.auth.onAuthStateChange((event, session) {
+      if (session != null && mounted) {
+        widget.onSuccess(session);
+      }
+    });
+  }
 
   @override
   void dispose() {
     _email.dispose();
+    _gotrueSubscription.data?.unsubscribe();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return Form(
-      autovalidateMode: AutovalidateMode.onUserInteraction,
       key: _formKey,
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -50,13 +70,13 @@ class _SupaMagicAuthState extends State<SupaMagicAuth> {
             },
             decoration: const InputDecoration(
               prefixIcon: Icon(Icons.email),
-              hintText: 'Enter your email',
+              label: Text('Enter your email'),
             ),
             controller: _email,
           ),
           spacer(16),
           ElevatedButton(
-            child: (isLoading)
+            child: (_isLoading)
                 ? const SizedBox(
                     height: 16,
                     width: 16,
@@ -74,30 +94,35 @@ class _SupaMagicAuthState extends State<SupaMagicAuth> {
                 return;
               }
               setState(() {
-                isLoading = true;
+                _isLoading = true;
               });
               try {
-                final result = await _supaAuth.createNewPasswordlessUser(
-                    _email.text,
-                    redirectUrl: widget.redirectUrl);
-                widget.onSuccess?.call(result);
+                await supaClient.auth.signIn(
+                  email: _email.text,
+                  options: AuthOptions(
+                    redirectTo: widget.redirectUrl,
+                  ),
+                );
                 if (mounted) {
-                  successSnackBar(context, 'Created passwordless user !');
+                  context.showSnackBar('Check your email inbox!');
                 }
               } on GoTrueException catch (error) {
-                if (widget.onError == null ||
-                    widget.onError?.call(error) == false) {
-                  await warningSnackBar(context, error.message);
+                if (widget.onError == null) {
+                  context.showErrorSnackBar(error.message);
+                } else {
+                  widget.onError?.call(error);
                 }
               } catch (error) {
-                await warningSnackBar(
-                    context, 'Unexpected error has occurred: $error');
+                if (widget.onError == null) {
+                  context.showErrorSnackBar(
+                      'Unexpected error has occurred: $error');
+                } else {
+                  widget.onError?.call(error);
+                }
               }
-              if (mounted) {
-                setState(() {
-                  isLoading = false;
-                });
-              }
+              setState(() {
+                _isLoading = false;
+              });
             },
           ),
           spacer(10),