+++ lib/src/components/supa_send_email.dart
@@ -1,17 +1,27 @@
 import 'package:email_validator/email_validator.dart';
 import 'package:flutter/material.dart';
 import 'package:supabase_auth_ui/src/utils/constants.dart';
-import 'package:supabase_auth_ui/src/utils/supabase_auth_ui.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
+/// UI component to create password reset form
 class SupaSendEmail extends StatefulWidget {
+  /// `redirectUrl` to be passed to the `.signIn()` or `signUp()` methods
+  ///
+  /// Typically used to pass a DeepLink
   final String? redirectUrl;
-  final void Function(GotrueJsonResponse response)? onSuccess;
-  final bool Function(GoTrueException error)? onError;
 
-  const SupaSendEmail(
-      {Key? key, this.redirectUrl, this.onSuccess, this.onError})
-      : super(key: key);
+  /// Method to be called when the auth action is success
+  final void Function(GotrueJsonResponse response) onSuccess;
+
+  /// Method to be called when the auth action threw an excepction
+  final void Function(Object error)? onError;
+
+  const SupaSendEmail({
+    Key? key,
+    this.redirectUrl,
+    required this.onSuccess,
+    this.onError,
+  }) : super(key: key);
 
   @override
   State<SupaSendEmail> createState() => _SupaSendEmailState();
@@ -21,9 +31,7 @@ class _SupaSendEmailState extends State<SupaSendEmail> {
   final _formKey = GlobalKey<FormState>();
   final _email = TextEditingController();
 
-  SupabaseAuthUi supaAuth = SupabaseAuthUi();
-
-  bool isLoading = false;
+  bool _isLoading = false;
 
   @override
   void dispose() {
@@ -34,7 +42,6 @@ class _SupaSendEmailState extends State<SupaSendEmail> {
   @override
   Widget build(BuildContext context) {
     return Form(
-      autovalidateMode: AutovalidateMode.onUserInteraction,
       key: _formKey,
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -50,13 +57,13 @@ class _SupaSendEmailState extends State<SupaSendEmail> {
             },
             decoration: const InputDecoration(
               prefixIcon: Icon(Icons.email),
-              hintText: 'Enter your email',
+              label: Text('Enter your email'),
             ),
             controller: _email,
           ),
           spacer(16),
           ElevatedButton(
-            child: (isLoading)
+            child: (_isLoading)
                 ? const SizedBox(
                     height: 16,
                     width: 16,
@@ -74,27 +81,34 @@ class _SupaSendEmailState extends State<SupaSendEmail> {
                 return;
               }
               setState(() {
-                isLoading = true;
+                _isLoading = true;
               });
               try {
-                final result = await supaAuth.sendResetPasswordEmail(
-                    _email.text, widget.redirectUrl);
-                widget.onSuccess?.call(result);
-                if (mounted) {
-                  successSnackBar(context, 'Email successfully sent !');
-                }
+                final response =
+                    await supaClient.auth.api.resetPasswordForEmail(
+                  _email.text,
+                  options: AuthOptions(
+                    redirectTo: widget.redirectUrl,
+                  ),
+                );
+                widget.onSuccess.call(response);
               } on GoTrueException catch (error) {
-                if (widget.onError == null ||
-                    widget.onError?.call(error) == false) {
-                  await warningSnackBar(context, error.message);
+                if (widget.onError == null) {
+                  context.showErrorSnackBar(error.message);
+                } else {
+                  widget.onError?.call(error);
                 }
               } catch (error) {
-                await warningSnackBar(
-                    context, 'Unexpected error has occurred: $error');
+                if (widget.onError == null) {
+                  context.showErrorSnackBar(
+                      'Unexpected error has occurred: $error');
+                } else {
+                  widget.onError?.call(error);
+                }
               }
               if (mounted) {
                 setState(() {
-                  isLoading = false;
+                  _isLoading = false;
                 });
               }
             },