+++ lib/src/utils/constants.dart
@@ -1,29 +1,41 @@
 import 'package:flutter/material.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
 
-Future<void> successSnackBar(BuildContext context, String succMsg) async {
-  final snackBar = SnackBar(
-    content: Text(style: const TextStyle(color: Colors.green), succMsg),
-    action: SnackBarAction(
-      label: 'Ok',
-      onPressed: () {},
-    ),
-  );
-  ScaffoldMessenger.of(context).showSnackBar(snackBar);
-}
-
-Future<void> warningSnackBar(BuildContext context, String errMsg) async {
-  final snackBar = SnackBar(
-    content: Text(style: const TextStyle(color: Colors.redAccent), errMsg),
-    action: SnackBarAction(
-      label: 'Ok',
-      onPressed: () {},
-    ),
-  );
-  ScaffoldMessenger.of(context).showSnackBar(snackBar);
-}
+final supaClient = Supabase.instance.client;
 
 SizedBox spacer(double height) {
   return SizedBox(
     height: height,
   );
+}
+
+/// Set of extension methods to easily display a snackbar
+extension ShowSnackBar on BuildContext {
+  /// Displays a basic snackbar
+  void showSnackBar(
+    String message, {
+    Color? backgroundColor,
+    String? actionLabel,
+  }) {
+    ScaffoldMessenger.of(this).showSnackBar(SnackBar(
+      content: Text(message),
+      backgroundColor: backgroundColor,
+      action: SnackBarAction(
+        label: actionLabel ?? 'ok',
+        onPressed: () {},
+      ),
+    ));
+  }
+
+  /// Displays a red snackbar indicating error
+  void showErrorSnackBar(
+    String message, {
+    String? actionLabel,
+  }) {
+    showSnackBar(
+      message,
+      backgroundColor: Theme.of(this).errorColor,
+      actionLabel: actionLabel,
+    );
+  }
 }