+++ lib/src/utils/supa_auth_action.dart
@@ -1,0 +1,1 @@
+enum SupaAuthAction { signIn, signUp }